requirements.txt
# Locked dependencies for AI Mechanic - generated from working environment
# This prevents pip backtracking during Docker builds

# Core web framework
fastapi==0.108.0
uvicorn==0.25.0
starlette==0.32.0.post1

# HTTP clients and server components  
httpx==0.25.2
httpcore==1.0.9
h11==0.16.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
anyio==4.9.0
sniffio==1.3.1

# OpenAI integration
openai==1.12.0

# Database connections
supabase==2.3.4
psycopg2-binary==2.9.9
pgvector==0.2.5
postgrest==0.15.1
gotrue==2.9.1
storage3==0.7.7
supafunc==0.3.3
realtime==1.0.6

# Document processing
PyMuPDF==1.23.14
PyMuPDFb==1.23.9

# Scientific computing - exact working versions
numpy==1.24.3
scipy==1.13.1
scikit-learn==1.6.1
joblib==1.5.1
threadpoolctl==3.6.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
tiktoken==0.5.2
tqdm==4.67.1

# Type checking and validation
pydantic==2.11.5
pydantic_core==2.33.2
annotated-types==0.7.0
typing_extensions==4.13.2
typing-inspection==0.4.1

# Other dependencies
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
idna==3.10
packaging==25.0
python-dateutil==2.9.0.post0
regex==2024.11.6
StrEnum==0.4.15
urllib3==1.26.6
websockets==12.0